调试小程序注册后台账号功能

+++ backend/common/src/main/java/com/zxiaosi/common/entity/Role.java
@@ -34,7 +34,7 @@ public class Role implements Serializable {
     /**
      * 是否删除
      */
-    private Short isDeleted;
+    private Integer isDeleted;
 
     /**
      * 创建时间
